fix: color moved to top of reexport
