+++ packages/ui/src/avatar-button/AvatarButton.tsx
@@ -15,7 +15,7 @@ import {
   Ref,
   RefAttributes,
 } from 'react';
-import { Color } from '../theme/Color';
+import { Color } from '../color/Color';
 import { SuperDispatchTheme } from '../theme/SuperDispatchTheme';
 
 export type AvatarButtonClassKey =
