+++ packages/ui/src/card-button/CardButton.tsx
@@ -12,7 +12,7 @@ import {
   ReactNode,
   RefAttributes,
 } from 'react';
-import { Color } from '../theme/Color';
+import { Color } from '../color/Color';
 
 export type CardButtonClassKey =
   | 'root'
