+++ packages/ui/src/description-list/DescriptionList.tsx
@@ -2,8 +2,8 @@ import { SvgIcon, Typography, TypographyProps } from '@material-ui/core';
 import { CSSProperties, makeStyles } from '@material-ui/styles';
 import clsx from 'clsx';
 import { forwardRef, ReactNode } from 'react';
+import { Color } from '../color/Color';
 import { OverflowText, OverflowTextProps } from '../overflow-text/OverflowText';
-import { Color } from '../theme/Color';
 import { SuperDispatchTheme } from '../theme/SuperDispatchTheme';
 import { isEmptyReactNode } from '../utils/isEmptyReactNode';
 import { useUID } from '../utils/useUID';
