+++ packages/ui/src/drawer/DrawerTitle.tsx
@@ -8,7 +8,7 @@ import {
 import { makeStyles } from '@material-ui/styles';
 import clsx from 'clsx';
 import { forwardRef, HTMLAttributes, ReactNode } from 'react';
-import { Color } from '../theme/Color';
+import { Color } from '../color/Color';
 import { SuperDispatchTheme } from '../theme/SuperDispatchTheme';
 import { VisibilityObserver } from '../utils/VisibilityObserver';
