+++ packages/ui/src/index.ts
@@ -5,6 +5,7 @@ export * from './button/Button';
 export * from './card-button/CardButton';
 export * from './checkbox/CheckboxField';
 export * from './checkbox/CheckboxGroudField';
+export * from './color/Color';
 export * from './columns/Column';
 export * from './columns/Columns';
 export * from './description-list/DescriptionList';
@@ -25,7 +26,6 @@ export * from './props/CollapseProp';
 export * from './props/ResponsiveProp';
 export * from './props/SpaceProp';
 export * from './radio/RadioField';
-/* prettier-ignore eslint-disable */
 export * from './radio/RadioFieldCard';
 export * from './radio/RadioGroupField';
 export * from './responsive/CollapseBreakpoint';
@@ -36,7 +36,6 @@ export * from './snackbar/SnackbarContent';
 export * from './snackbar/SnackbarStack';
 export * from './stack/Stack';
 export * from './tag/Tag';
-export * from './theme/Color';
 export * from './theme/SuperDispatchTheme';
 export * from './theme/ThemeProvider';
 export * from './tiles/Tiles';
