+++ packages/ui/src/overflow-text/OverflowText.tsx
@@ -12,7 +12,7 @@ import {
   ForwardRefExoticComponent,
   useState,
 } from 'react';
-import { Color } from '../theme/Color';
+import { Color } from '../color/Color';
 import { SuperDispatchTheme } from '../theme/SuperDispatchTheme';
 import { VisibilityObserver } from '../utils/VisibilityObserver';
