+++ packages/ui/src/radio/RadioOverrides.tsx
@@ -1,5 +1,5 @@
 import { SvgIcon } from '@material-ui/core';
-import { Color } from '../theme/Color';
+import { Color } from '../color/Color';
 import { SuperDispatchTheme } from '../theme/SuperDispatchTheme';
 
 export function overrideRadio(theme: SuperDispatchTheme): void {
