+++ packages/ui/src/svg-icon/SvgIconOverrides.ts
@@ -1,4 +1,4 @@
-import { Color } from '../theme/Color';
+import { Color } from '../color/Color';
 import { SuperDispatchTheme } from '../theme/SuperDispatchTheme';
 
 export function overrideSvgIcon(theme: SuperDispatchTheme): void {
